Add MetaList data structure
A MetaList<...> is a class that represents a list of types (stored
as a parameter pack) that can be manipulated at compile time.
Because all MetaList operations occur at compile time, all of the
test cases use static_assert(). If the test case source file compiles,
then all of the tests passed.