Overhaul build system
Overhaul the build system to better allow Estoppel to be used as an
external dependency in independent applications. Introduces the
concept of a target platform for a build, with the target platform
providing the machine compile flags and other critical build-time
information. Also adds binary, Intel hex, and size information outputs
to device executables by default.