Add Bitfield class
A Bitfield is intended to be a member of a union, and allows
a relatively portable means for directly reading and
manipulating bit fields within a larger word. Fields are represented
as integral values or as enumerations. When enumerations
are used, automatic selection of the underlying storage type
is provided.