Update assertion handling
When building for a host target (as indicated by the
ESTP_HOST #define) call std::terminate in the default
assertion handler. This ensures that test cases that fail
are properly detected.

Also, check for previous definition of ASSERT() and safely redfine
if necessary. The Nordic SDK defines ASSERT() and provides an
assertion handling system. This change allows the Estoppel assert
system to detect this situation and take overcontrol of ASSERT().

A compile time message is generated if ASSERT() is redefined.