Add Accumulator data structure
An accumulator is essentially a stack, and data can only be
pushed or popped from the back. However, unlike a stack a
FIFO read-write view (as a Span) of all of the data can be
retrieved. This allows an Accumulator to be used to buffer
data and then retrieve the entire buffer at once.